ci: test architecture integer type sizes with new RNG

+++ src/common.jl
@@ -30,7 +30,7 @@ const ROOT_3_OVER_3 = 0.577350269189626
 ### Random number generation
 
 struct RandomState
-    seed::UInt64
+    seed::Int
     rng::Xoroshiro128Star
 end
